fix: Handle error offset correctly
In the original implementation, we track an absolute
error (`lastBoldLossError_Offset`), based on some amount of total Bold
deposits, but afterwards we add it to a fraction with a smaller amount
of total Bold deposits, amplifying the effect of this error recovery
too much.

This causes that over a sequence of liquidations, the ratio between
the initial P and the final P is bigger than the ratio between the
initial total amount of deposits and its final value. This is fine in
one step if it’s compensating for a previous error, but not on the
long run. If we compare to an initial state where there was no error,
the ratio between P’s should be always less or equal than the ratio
between total deposits. Of course theoretically should be exactly the
same, but in practice, we must handle the rounding errors that way to
prevent that sum of yield rewards (which use P ratios) grows bigger
than the total amount of owed rewards (causing an underflow in the
subtraction if all users would try to claim).